+++ mods/Mining/mod/data-updates.lua
@@ -1,3 +1,2 @@
 
 require("prototypes.recipe.recipe_changes")
-require("prototypes.smithables")