+++ mods/Smithing/mod/data-updates.lua
@@ -1,2 +1,3 @@
 
 require("prototypes.furnace-updates")
+require("prototypes.smithables")