+++ mods/Smithing/mod/data.lua
@@ -7,10 +7,12 @@ require("prototypes.item.cables")
 
 require("prototypes.anvil")
 
+require("prototypes.recipe.armour_recipe")
 require("prototypes.recipe.bar_recipe")
 require("prototypes.recipe.intermediate_recipe")
 require("prototypes.recipe.pickaxe_recipe")
 
+require("prototypes.technology.armour_technology")
 require("prototypes.technology.processing_technology")
 require("prototypes.technology.pickaxe_technology")
 require("prototypes.technology.smelting_technology")