+++ mods/Smithing/mod/data/armour_data.lua
@@ -1,0 +1,52 @@
+return {
+  ["bronze"] = {
+    order = 1,
+    shield = "kite",
+    base_speed = 1
+  },
+  ["iron"] = {
+    order = 10,
+    shield = "kite",
+    base_speed = 2
+  },
+  ["steel"] = {
+    order = 20,
+    shield = "kite",
+    base_speed = 3
+  },
+  ["mithril"] = {
+    order = 30,
+    shield = "kite",
+    base_speed = 5
+  },
+  ["adamant"] = {
+    order = 40,
+    shield = "kite",
+    base_speed = 7
+  },
+  ["rune"] = {
+    order = 50,
+    shield = "kite",
+    base_speed = 11
+  },
+  ["orikalkum"] = {
+    order = 60,
+    shield = "kite",
+    base_speed = 13
+  },
+  ["necronium"] = {
+    order = 70,
+    shield = "kite",
+    base_speed = 17
+  },
+  ["bane"] = {
+    order = 80,
+    shield = "square-",
+    base_speed = 19
+  },
+  ["elder-rune"] = {
+    order = 90,
+    shield = "round-",
+    base_speed = 23
+  }
+}