+++ mods/Smithing/mod/prototypes/smithables.lua
@@ -1,4 +1,5 @@
 require("__RsSmithing__.utility.smithing_util")
+local armours = require("__RsSmithing__.data.armour_data")
 
 make_smithable("rs-bronze-cable", "rs-bronze-processing")
 make_smithable("rs-adamant-cable", "rs-adamant-processing")
@@ -14,4 +15,15 @@ local mats = {"bronze", "iron", "steel", "mithril", "adamant", "rune", "orikalku
 
 for _, mat in pairs(mats) do
   make_smithable("rs-"..mat.."-pickaxe", "rs-"..mat.."-pickaxe")
+end
+
+
+for mat, data in pairs(armours) do
+  local tech = "rs-"..mat.."-armour"
+  make_smithable("rs-"..mat.."-gauntlets", tech)
+  make_smithable("rs-"..mat.."-armoured-boots", tech)
+  make_smithable("rs-"..mat.."-full-helm", tech)
+  make_smithable("rs-"..mat.."-"..data.shield.."shield", tech)
+  make_smithable("rs-"..mat.."-platelegs", tech)
+  make_smithable("rs-"..mat.."-platebody", tech)
 end