+++ mods/Smithing/mod/prototypes/technology/smelting_technology.lua
@@ -31,7 +31,7 @@ local iron_bar_tech = {
   levels = {
     ["smithing"] = 10
   },
-  icon = "__RsMining__/graphics/technology/iron_bar.png",
+  icon = "__RsSmithing__/graphics/technology/iron_bar.png",
   icon_size = 685,
   localised_name = {"item-name.rs-iron-bar"}
 }
@@ -50,7 +50,7 @@ local steel_bar_tech = {
   levels = {
     ["smithing"] = 20
   },
-  icon = "__RsMining__/graphics/technology/steel_bar.png",
+  icon = "__RsSmithing__/graphics/technology/steel_bar.png",
   icon_size = 685,
   extra_prerequisites = {"rs-iron-bar"},
   localised_name = {"item-name.rs-steel-bar"}
@@ -69,7 +69,7 @@ local mithril_bar_tech = {
   levels = {
     ["smithing"] = 30
   },
-  icon = "__RsMining__/graphics/technology/mithril_bar.png",
+  icon = "__RsSmithing__/graphics/technology/mithril_bar.png",
   icon_size = 685,
   extra_prerequisites = {"rs-steel-bar"},
   localised_name = {"item-name.rs-mithril-bar"}
@@ -88,7 +88,7 @@ local adamant_bar_tech = {
   levels = {
     ["smithing"] = 40
   },
-  icon = "__RsMining__/graphics/technology/adamant_bar.png",
+  icon = "__RsSmithing__/graphics/technology/adamant_bar.png",
   icon_size = 685,
   extra_prerequisites = {"rs-mithril-bar"},
   localised_name = {"item-name.rs-adamant-bar"}
@@ -108,7 +108,7 @@ local rune_bar_tech = {
   levels = {
     ["smithing"] = 50
   },
-  icon = "__RsMining__/graphics/technology/rune_bar.png",
+  icon = "__RsSmithing__/graphics/technology/rune_bar.png",
   icon_size = 685,
   extra_prerequisites = {"rs-adamant-bar"},
   localised_name = {"item-name.rs-rune-bar"}
@@ -127,7 +127,7 @@ local orikalkum_bar_tech = {
   levels = {
     ["smithing"] = 60
   },
-  icon = "__RsMining__/graphics/technology/orikalkum_bar.png",
+  icon = "__RsSmithing__/graphics/technology/orikalkum_bar.png",
   icon_size = 685,
   extra_prerequisites = {"rs-rune-bar"},
   localised_name = {"item-name.rs-orikalkum-bar"}
@@ -147,7 +147,7 @@ local necronium_bar_tech = {
   levels = {
     ["smithing"] = 70
   },
-  icon = "__RsMining__/graphics/technology/necronium_bar.png",
+  icon = "__RsSmithing__/graphics/technology/necronium_bar.png",
   icon_size = 685,
   extra_prerequisites = {"rs-orikalkum-bar"},
   localised_name = {"item-name.rs-necronium-bar"}
@@ -166,7 +166,7 @@ local bane_bar_tech = {
   levels = {
     ["smithing"] = 80
   },
-  icon = "__RsMining__/graphics/technology/bane_bar.png",
+  icon = "__RsSmithing__/graphics/technology/bane_bar.png",
   icon_size = 685,
   extra_prerequisites = {"rs-necronium-bar"},
   localised_name = {"item-name.rs-necronium-bar"}
@@ -185,7 +185,7 @@ local elder_rune_bar_tech = {
   levels = {
     ["smithing"] = 90
   },
-  icon = "__RsMining__/graphics/technology/elder_rune_bar.png",
+  icon = "__RsSmithing__/graphics/technology/elder_rune_bar.png",
   icon_size = 685,
   extra_prerequisites = {"rs-bane-bar"},
   localised_name = {"item-name.rs-elder-rune-bar"}